requirements.txt
click
python-dotenv
tweepy
